Create storyUtils and filters buttons in the Form comp
