+++ src/forms/Form/Form.stories.tsx
@@ -9,6 +9,7 @@ import { useForm, formActions } from "@root/forms/Form";
 import { useImageVideoLinkDocumentBrowsing, imageVideoSrc } from "@root/forms/FormFieldImageVideoLinkDocumentBrowsing/ImageVideoLinkDocumentBrowsingUtils";
 import { validateEmail, validateSlow, required, validateNumber, validateURL } from "./validators";
 import { menuOptions } from "../MenuFormFieldCard/MenuFormFieldUtils";
+import { onCancel, renderButtons } from "@root/utils/storyUtils";
 
 // Components
 import Form from "./Form";
@@ -121,34 +122,6 @@ const getSelected = async (selectedOptions) => {
 		externalOptions.find(o => o.value === selectedOption)
 	);
 }
-
-const onCancel = () => {
-	alert("Cancelling form, going back to previous site");
-};
-
-const onSubmit = async (dispatch) => {
-	const { valid, data } = await dispatch(formActions.submitForm());
-	if (!valid) return;
-
-	alert("Form submitted with the following data: " + JSON.stringify(data, null, " "));
-};
-
-const renderButtons = (dispatch, show = { showCancel: true, showSave: true }): ButtonProps[] => [
-	{
-		label: "Cancel",
-		onClick: onCancel,
-		color: "gray",
-		variant: "outlined",
-		show: show.showCancel,
-	},
-	{
-		label: "Save",
-		onClick: () => onSubmit(dispatch),
-		color: "yellow",
-		variant: "contained",
-		show: show.showSave
-	},
-];
 
 export const Playground = (): ReactElement => {
 	const [loadReady, setLoadReady] = useState(false);
