+++ src/forms/FormFieldAddress/Address.stories.tsx
@@ -1,9 +1,9 @@
 import * as React from "react";
 import { ReactElement, useMemo } from "react";
 import { boolean, text, withKnobs } from "@storybook/addon-knobs";
-import { ButtonProps } from "@root/components/Button";
+import { onCancel, renderButtons } from "@root/utils/storyUtils";
 import { FieldDef } from "../../components/Field";
-import { useForm, formActions } from "../Form";
+import { useForm } from "../Form";
 
 // Components
 import Form from "../Form/Form";
@@ -12,32 +12,6 @@ export default {
 	title: "FormFields/FormFieldAddress",
 	decorators: [withKnobs],
 };
-
-const onCancel = () => {
-	alert("Cancelling form, going back to previous site");
-};
-
-const onSubmit = async (dispatch) => {
-	const { valid, data } = await dispatch(formActions.submitForm());
-	if (!valid) return;
-
-	alert("Form submitted with the following data: " + JSON.stringify(data, null, " "));
-};
-
-const renderButtons = (dispatch): ButtonProps[] => [
-	{
-		label: "Save",
-		onClick: () => onSubmit(dispatch),
-		color: "yellow",
-		variant: "contained",
-	},
-	{
-		label: "Cancel",
-		onClick: onCancel,
-		color: "gray",
-		variant: "outlined",
-	},
-];
 
 export const Playground = (): ReactElement => {
 	const { state, dispatch, registerFields } = useForm();
