+++ src/forms/FormFieldChipSingleSelect/FormFieldChipSingleSelect.stories.tsx
@@ -4,8 +4,8 @@ import { boolean, text, withKnobs } from "@storybook/addon-knobs";
 import { Meta } from "@storybook/addon-docs/blocks";
 import { FormFieldChipSingleSelectDef } from ".";
 import { FieldDef } from "@root/components/Field";
-import { useForm, formActions } from "../Form";
-import { ButtonProps } from "@root/components/Button";
+import { useForm } from "../Form";
+import { onCancel, renderButtons } from "@root/utils/storyUtils";
 
 // Components
 import Form from "../Form/Form";
@@ -14,32 +14,6 @@ export default {
 	title: "FormFields/FormFieldChipSingleSelect",
 	decorators: [withKnobs],
 } as Meta;
-
-const onCancel = () => {
-	alert("Cancelling form, going back to previous site");
-};
-
-const onSubmit = async (dispatch) => {
-	const { valid, data } = await dispatch(formActions.submitForm());
-	if (!valid) return;
-
-	alert("Form submitted with the following data: " + JSON.stringify(data, null, " "));
-};
-
-const renderButtons = (dispatch): ButtonProps[] => [
-	{
-		label: "Save",
-		onClick: () => onSubmit(dispatch),
-		color: "yellow",
-		variant: "contained",
-	},
-	{
-		label: "Cancel",
-		onClick: onCancel,
-		color: "gray",
-		variant: "outlined",
-	},
-];
 
 export const Playground = (): ReactElement => {
 	const { state, dispatch, registerFields } = useForm();
