+++ src/forms/FormFieldMapCoordinates/MapCoordinates.stories.tsx
@@ -3,45 +3,19 @@ import { ReactElement, useMemo } from "react";
 import { boolean, withKnobs, object, text } from "@storybook/addon-knobs";
 import { FieldDef } from "@root/components/Field";
 import { MapCoordinatesDef } from ".";
-import { ButtonProps } from "@root/components/Button";
+import { onCancel, renderButtons } from "@root/utils/storyUtils";
 
 // Components
 import Form from "../Form/Form";
 
 // Utils
-import { useForm, formActions } from "../Form";
+import { useForm } from "../Form";
 import { address, defaultMapPosition } from "./MapCoordinatesUtils";
 
 export default {
 	title: "FormFields/FormFieldMapCoordinates",
 	decorators: [withKnobs],
 };
-
-const onCancel = () => {
-	alert("Cancelling form, going back to previous site");
-};
-
-const onSubmit = async (dispatch) => {
-	const { valid, data } = await dispatch(formActions.submitForm());
-	if (!valid) return;
-
-	alert("Form submitted with the following data: " + JSON.stringify(data, null, " "));
-};
-
-const renderButtons = (dispatch): ButtonProps[] => [
-	{
-		label: "Save",
-		onClick: () => onSubmit(dispatch),
-		color: "yellow",
-		variant: "contained",
-	},
-	{
-		label: "Cancel",
-		onClick: onCancel,
-		color: "gray",
-		variant: "outlined",
-	},
-];
 
 export const Playground = (): ReactElement => {
 	const { state, dispatch, registerFields } = useForm();
