+++ src/forms/FormFieldPhoneSelectionDropdown/FormFieldPhoneSelectionDropdown.stories.tsx
@@ -2,9 +2,9 @@ import * as React from "react";
 import { ReactElement, useMemo } from "react";
 import { boolean, withKnobs, text } from "@storybook/addon-knobs";
 import { PhoneSelectionDef } from "./FormFieldPhoneSelectionDropdownTypes";
-import { ButtonProps } from "@root/components/Button";
 import { FieldDef } from "@root/components/Field";
-import { useForm, formActions } from "../Form";
+import { useForm } from "../Form";
+import { onCancel, renderButtons } from "@root/utils/storyUtils";
 
 // Components
 import Form from "../Form/Form";
@@ -13,32 +13,6 @@ export default {
 	title: "FormFields/FormFieldPhoneSelectionDropdown",
 	decorators: [withKnobs],
 };
-
-const onCancel = () => {
-	alert("Cancelling form, going back to previous site");
-};
-
-const onSubmit = async (dispatch) => {
-	const { valid, data } = await dispatch(formActions.submitForm());
-	if (!valid) return;
-
-	alert("Form submitted with the following data: " + JSON.stringify(data, null, " "));
-};
-
-const renderButtons = (dispatch): ButtonProps[] => [
-	{
-		label: "Save",
-		onClick: () => onSubmit(dispatch),
-		color: "yellow",
-		variant: "contained",
-	},
-	{
-		label: "Cancel",
-		onClick: onCancel,
-		color: "gray",
-		variant: "outlined",
-	},
-];
 
 export const Playground = () : ReactElement => {
 	const { state, dispatch, registerFields } = useForm();
