+++ src/forms/FormFieldRadio/FormFieldRadio.stories.tsx
@@ -2,10 +2,10 @@ import * as React from "react";
 import { ReactElement, useMemo } from "react";
 import { boolean, withKnobs, text } from "@storybook/addon-knobs";
 import { Meta } from "@storybook/addon-docs/blocks";
-import { useForm, formActions } from "../Form";
-import { ButtonProps } from "@root/components/Button";
+import { useForm } from "../Form";
 import { FormFieldRadioDef } from ".";
 import { FieldDef } from "@root/components/Field";
+import { onCancel, renderButtons } from "@root/utils/storyUtils";
 
 // Components
 import Form from "@root/forms/Form/Form";
@@ -27,32 +27,6 @@ const options = [
 	{
 		label: "Label 3",
 		value: "label_3",
-	},
-];
-
-const onCancel = () => {
-	alert("Cancelling form, going back to previous site");
-};
-
-const onSubmit = async (dispatch) => {
-	const { valid, data } = await dispatch(formActions.submitForm());
-	if (!valid) return;
-
-	alert("Form submitted with the following data: " + JSON.stringify(data, null, " "));
-};
-
-const renderButtons = (dispatch): ButtonProps[] => [
-	{
-		label: "Save",
-		onClick: () => onSubmit(dispatch),
-		color: "yellow",
-		variant: "contained",
-	},
-	{
-		label: "Cancel",
-		onClick: onCancel,
-		color: "gray",
-		variant: "outlined",
 	},
 ];
