+++ src/forms/FormFieldTable/Table.stories.tsx
@@ -8,44 +8,18 @@ import { FieldDef } from "@root/components/Field";
 import Form from "../Form/Form";
 
 // Utils
-import { useForm, formActions } from "../Form";
+import { useForm } from "../Form";
 import {
 	headers,
 	deleteTableRow,
 	useTable,
 } from "@root/forms/FormFieldTable/tableUtils";
-import { ButtonProps } from "@root/components/Button";
+import { onCancel, renderButtons } from "@root/utils/storyUtils";
 
 export default {
 	title: "FormFields/FormFieldTable",
 	decorators: [withKnobs],
 };
-
-const onCancel = () => {
-	alert("Cancelling form, going back to previous site");
-};
-
-const onSubmit = async (dispatch) => {
-	const { valid, data } = await dispatch(formActions.submitForm());
-	if (!valid) return;
-
-	alert("Form submitted with the following data: " + JSON.stringify(data, null, " "));
-};
-
-const renderButtons = (dispatch): ButtonProps[] => [
-	{
-		label: "Save",
-		onClick: () => onSubmit(dispatch),
-		color: "yellow",
-		variant: "contained",
-	},
-	{
-		label: "Cancel",
-		onClick: onCancel,
-		color: "gray",
-		variant: "outlined",
-	},
-];
 
 export const Playground = (): ReactElement => {
 	const { state, dispatch, registerFields } = useForm();
