+++ src/forms/FormFieldText/FormFieldText.stories.tsx
@@ -7,10 +7,10 @@ import {
 	number,
 	select,
 } from "@storybook/addon-knobs";
-import { ButtonProps } from "@root/components/Button";
 import { TextFieldDef } from ".";
 import { FieldDef } from "@root/components/Field";
-import { useForm, formActions } from "../Form";
+import { useForm } from "../Form";
+import { onCancel, renderButtons } from "@root/utils/storyUtils";
 
 // Components
 import AccountCircle from "@mui/icons-material/AccountCircle";
@@ -20,32 +20,6 @@ export default {
 	title: "FormFields/FormFieldText",
 	decorators: [withKnobs],
 };
-
-const onCancel = () => {
-	alert("Cancelling form, going back to previous site");
-};
-
-const onSubmit = async (dispatch) => {
-	const { valid, data } = await dispatch(formActions.submitForm());
-	if (!valid) return;
-
-	alert("Form submitted with the following data: " + JSON.stringify(data, null, " "));
-};
-
-const renderButtons = (dispatch): ButtonProps[] => [
-	{
-		label: "Save",
-		onClick: () => onSubmit(dispatch),
-		color: "yellow",
-		variant: "contained",
-	},
-	{
-		label: "Cancel",
-		onClick: onCancel,
-		color: "gray",
-		variant: "outlined",
-	},
-];
 
 export const Playground = (): ReactElement => {
 	const {	state, dispatch, registerFields	} = useForm();
@@ -107,6 +81,7 @@ export const Playground = (): ReactElement => {
 		<>
 			<pre>{JSON.stringify(state, null, "  ")}</pre>
 			<Form
+				buttons={renderButtons(dispatch)}
 				title={text("Title", "Form Title")}
 				description={text("Description", "This is a description example")}
 				state={state}
