+++ src/forms/FormFieldTextArea/FormFieldTextArea.stories.tsx
@@ -9,8 +9,8 @@ import {
 } from "@storybook/addon-knobs";
 import { TextAreaDef } from ".";
 import { FieldDef } from "@root/components/Field";
-import { useForm, formActions } from "../Form";
-import { ButtonProps } from "@root/components/Button";
+import { useForm } from "../Form";
+import { onCancel, renderButtons } from "@root/utils/storyUtils";
 
 // Components
 import Form from "../Form/Form";
@@ -19,32 +19,6 @@ export default {
 	title: "FormFields/FormFieldTextArea",
 	decorators: [withKnobs],
 }
-
-const onCancel = () => {
-	alert("Cancelling form, going back to previous site");
-};
-
-const onSubmit = async (dispatch) => {
-	const { valid, data } = await dispatch(formActions.submitForm());
-	if (!valid) return;
-
-	alert("Form submitted with the following data: " + JSON.stringify(data, null, " "));
-};
-
-const renderButtons = (dispatch): ButtonProps[] => [
-	{
-		label: "Save",
-		onClick: () => onSubmit(dispatch),
-		color: "yellow",
-		variant: "contained",
-	},
-	{
-		label: "Cancel",
-		onClick: onCancel,
-		color: "gray",
-		variant: "outlined",
-	},
-];
 
 export const Playground = (): ReactElement => {
 	const {
