+++ src/forms/FormFieldToggleSwitch/FormFieldToggleSwitch.stories.tsx
@@ -7,39 +7,13 @@ import { Meta } from "@storybook/addon-docs/blocks";
 import { FormFieldToggleSwitchDef } from ".";
 import Form from "../Form/Form";
 import { FieldDef } from "@root/components/Field";
-import { useForm, formActions } from "../Form";
-import { ButtonProps } from "@root/components/Button";
+import { useForm } from "../Form";
+import { onCancel, renderButtons } from "@root/utils/storyUtils";
 
 export default {
 	title: "FormFields/FormFieldToggleSwitch",
 	decorators: [withKnobs],
 } as Meta;
-
-const onSubmit = async (dispatch) => {
-	const { valid, data } = await dispatch(formActions.submitForm());
-	if (!valid) return;
-
-	alert("Form submitted with the following data: " + JSON.stringify(data, null, " "));
-};
-
-const onCancel = () => {
-	alert("Cancelling form, going back to previous site");
-};
-
-const renderButtons = (dispatch): ButtonProps[] => [
-	{
-		label: "Save",
-		onClick: () => onSubmit(dispatch),
-		color: "yellow",
-		variant: "contained",
-	},
-	{
-		label: "Cancel",
-		onClick: onCancel,
-		color: "gray",
-		variant: "outlined",
-	},
-];
 
 export const Playground = (): ReactElement => {
 	const { state, dispatch, registerFields } = useForm();
